Add the start of GPU image filtering
Here we have the notion of a generator that uses a gl program
to populate successive slices of a 3D texture.

TODO: implement multi-input, multi-output filters and also
chains of filters that get updated when the end of the
filter chain is requested (as in a render operation triggered
by a change of input variables).