WIP: add registration example
This example requires two image fields and a transform field.
A new generator class performs iterative updates of the
transform using a shader program.

TODO: generator.js needs to be updated to handle transformFields
as outputs.  This will require at least
- 3 component texture writing
- extracting slice iteration parameters from deformation grid datasets

And the big TODO is implementing a working registration algorithm.